Converter: remove `__destruct()` registration
This change removes the need for `wpdb` extensions (like in the Converter) to register their own shutdown methods.

This is necessary because the `__destruct()` method was removed in the parent `wpdb` class, so attempting to add it in the subclass will generate PHP errors.

See #WP21402.

Fixes #3508.

Props GDragoN, dd32.

In branches/2.6, for 2.6.10.

git-svn-id: https://bbpress.svn.wordpress.org/branches/2.6@7256 9866e705-20ec-0310-96e7-cbb4277adcfb