Roles: prevent notices when calling `bbp_add_forums_roles()`
This change adds sanity checks around the `$wp_roles` parameter of the `bbp_add_forums_roles()` function.

It will now attempt to initialize the `WP_Roles` class just-in-time, but only if not already doing the `wp_roles_init` action (to prevent recursion).

If after all of that `$wp_roles` is still not what it needs to be, this function will now silently bail instead of proceeding (to prevent debug notices & errors).

Fixes #3493.

Props naxoc.

In branches/2.6, for 2.6.10.

git-svn-id: https://bbpress.svn.wordpress.org/branches/2.6@7257 9866e705-20ec-0310-96e7-cbb4277adcfb