Admin: prevent broken list-table layouts.
This change adds CSS that targets forum/topic/reply list-table heads & foots.

It allows text – that is both unbreakable & wider than its parent th/td – to truncate with an ellipsis.

This is necessary because the default behavior is to continue to squish text down to an unreadable state, which happens frequently here due to the number of columns in the tables and the amount of content available inside of them to Keymasters & Moderators.

In branches/2.6, for 2.6.10.

git-svn-id: https://bbpress.svn.wordpress.org/branches/2.6@7261 9866e705-20ec-0310-96e7-cbb4277adcfb