Views: do not trap RSS feeds for unregistered views
This change adds a check inside of `bbp_request_feed_trap()` that only allows `bbp_display_topics_feed_rss2()` to happen if query-arguments exist for the slug of the view that is being requested.

Without this check in place, the view unintentionally includes all topics, because no additional arguments are passed into `bbp_has_topics()`.

Fixes #3544.

Props dd32.

In branches/2.6, for 2.6.10.

git-svn-id: https://bbpress.svn.wordpress.org/branches/2.6@7264 9866e705-20ec-0310-96e7-cbb4277adcfb