Search: remove `array_keys()` from
This change ensures that the correct public statuses are used by default inside of the `bbp_has_search_results()` function, and it does this by using the values from `bbp_get_public_topic_statuses()` and not the keys.

This fixes a regression causing search results to omit public statuses as part of their default arguments.

Props GDragoN.

Fixes #3607.



git-svn-id: https://bbpress.svn.wordpress.org/branches/2.6@7280 9866e705-20ec-0310-96e7-cbb4277adcfb